Add support for uploading Files from Data URL
When a File generated by `File.fromDataURL` is uploaded in test, the
guard clause in the Queue service (i.e. `if (fileBlob instanceof Blob)`)
rejects the `fileBlob`.

In Chrome, a File generated by `File.fromDataURL` will have the `Blob`
available under the `.blob` property.

The `upload` test helper wasn't declared with
`Ember.Test.registerAsyncHelper`, so it wasn't waiting on or blocking
other asynchrouns helpers.

Once declared with `registerAsyncHelper`, we'll want to import it once
and refer to it as if it were `click` or `visit`.

This commit adds documentation to the README about adding an `import`
statement to applications' test helpers.

The `import` statements in the example code were replaced by `/* global
upload */` magic comments. Ideally, applications will add an `upload` to
their linter configuration files as a globally accessible function.